done asking user for shape

+++ 05_Ask_User_for_Shape.py
@@ -1,7 +1,7 @@
 
 def num_check(question, type):
 
-    if type == int:
+    if type == float:
         error_type = "a number"
     else:
         error_type = "an integer"
@@ -29,70 +29,71 @@ what_shape = input('''What shape do you want to calculate the area and perimeter
   5. Calculate Trapezium
   6. Calculate Parallelogram
   7. Calculate Rhombus
-  Please enter a number associated with the shape: ''').strip().lower()
+  Please enter a number associated with the shape:
+  ''').strip().lower()
 
 if what_shape in ['1', '1.']:
-        r = num_check("Radius: ", int)
+        r = num_check("Radius: ", float)
         perimeter = 2 * math.pi * r
         area = math.pi * (r * r)
         print("Perimeter of circle = ", perimeter)
         print("Area of a circle = ", area)
 
 elif what_shape in ['2', '2.']:
-        b = num_check("Base:", int)
+        b = num_check("Base:", float)
         perimeter = b * 4
         area = b * b
         print("Perimeter of the Square = ", perimeter)
         print("Area of the Square = ", area)
 
 elif what_shape in ['3.', '3']:
-        b = num_check("Base: ", int)
-        h = num_check("Height: ", int)
+        b = num_check("Base: ", float)
+        h = num_check("Height: ", float)
         perimeter = 2 * (b + h)
         area = b * h
         print("Perimeter of rectangle = ", perimeter)
         print("Area of a rectangle = ", area)
 
 elif what_shape in ['4.', '4']:
-        s1 = num_check("Side 1: ", int)
-        s2 = num_check("Side 2:", int)
-        s3 = num_check("Side 3:", int)
+        s1 = num_check("Side 1: ", float)
+        s2 = num_check("Side 2:", float)
+        s3 = num_check("Side 3:", float)
         perimeter = s1 + s2 + s3
         print("Perimeter of triangle = ", perimeter)
-        b = num_check("Base: ", int)
-        h = num_check("Height: ", int)
+        b = num_check("Base: ", float)
+        h = num_check("Height: ", float)
         area = b * h
         print("Area of a triangle = ", area)
 
 elif what_shape in ['5', '5.']:
-        a = num_check("Top length: ", int)
-        b = num_check("Bottom length: ", int)
-        c = num_check("Side 1: ", int)
-        d =num_check("Side 2: ", int)
+        a = num_check("Top length: ", float)
+        b = num_check("Bottom length: ", float)
+        c = num_check("Side 1: ", float)
+        d =num_check("Side 2: ", float)
         perimeter = a + b + c + d
         print("Perimeter of a Trapezium = ", perimeter)
-        a = num_check("Top length: ", int)
-        b = num_check("Bottom length: ", int)
-        h = num_check("Height: ", int)
+        a = num_check("Top length: ", float)
+        b = num_check("Bottom length: ", float)
+        h = num_check("Height: ", float)
         area = ((a+b)/2)*h
         print("Area of a Trapezium = ", area)
 
 elif what_shape in ['6', '6.']:
-        b = num_check("Base: ", int)
-        s1 = num_check("Side 1", int)
+        b = num_check("Base: ", float)
+        s1 = num_check("Side 1", float)
         perimeter = 2 * (b * s1)
         print("Perimeter of a Parallelogram = ", perimeter)
-        b = num_check("Base: ", int)
-        h = num_check("Height: ", int)
+        b = num_check("Base: ", float)
+        h = num_check("Height: ", float)
         area = b * h
         print("Area of a Parallelogram = ", area)
 
 elif what_shape in ['7', '7.']:
-        s1 = num_check("Side: ", int)
+        s1 = num_check("Side: ", float)
         perimeter = 4 * s1
         print("Perimeter of a Rhombus = ", perimeter)
-        D1 = num_check("Diagonal 1: ", int)
-        D2 = num_check("Diagonal 2: ", int)
+        D1 = num_check("Diagonal 1: ", float)
+        D2 = num_check("Diagonal 2: ", float)
         area = (D1 * D2)/2
         print("Area of a Rhombus = ", area)
